Add new kolibri-launcher app
This new app is a simple Gio.Application that handles the
kolibri-channel uri scheme and launch the kolibri-gnome with the
channel-id parameter.

This new launcher could be useful to launch ChannelApplication directly
using `gio open kolibri-channel:3e464ee12f6a50a781cddf59147b48b1`. This
is not possible with the Pew app because by default it will launch the
GenericApplication so this utility launcher will just handle the scheme
and call the real app with the corresponding parameters.