Split desktop_launcher module
This patch splits the desktop_launcher python module in two different
modules, the "desktop" and the "launcher" modules.

The "desktop" module contains the whole gtk application and the
"launcher" one just contain the "kolibri-channel" gio scheme launcher.
In the future we can move all the gio schemes management to the launcher
application and pass all the information to the "desktop" application
using command line arguments.