Log in automatically in the Kolibri desktop app
Ask for a login token when the DBus proxy changes status to
STARTED. And invalidate the token (set the variable holding it back to
None) when the status is other than STARTED. The token expires, so it
is assummed that it will be used right away by the webview.

Add the login token to the conditions for considering the daemon
manager has started.

Encode the autologin URL inside the "next" query parameter of the
initialize URL. Since the autologin URL also has a "next" query
parameter, this can effectively be chained.

https://phabricator.endlessm.com/T32613