kolibri-gnome: Improve automatic provisioning
Update the patch for the deviceprovision API, to also allow null value
in the `language_id` field.  When this device setting is null, Kolibri
will use the browser default language.

Revert the code added in commit 22af40b6 for discovering the closest
language, and the added Python dependency.

Also: set null value in the other fields that allow it:
- The device name setting.
- The allow quest access settting.

https://phabricator.endlessm.com/T32824