Directly use MagicBus to manage Kolibri's state
With Kolibri 0.15, KolibriProcessBus provides a state machine to manage
starting and stopping Kolibri. With this new functionality, we can
remove most of the existing code in kolibri-daemon for stopping and
starting Kolibri, and instead populate KolibriServiceContext by using
MagicBus plugins. Instead of using LauncherProcess to communicate with a
a collection of Kolibri-related processes, KolibriServiceManager can use
a similar interface to communicate with a single DjangoProcess.

https://phabricator.endlessm.com/T32972