Roll version number back to 0.1
This app is not ready for consumption yet, so I don't think we should be
calling it 1.0.0. Furthermore, since this is an app and not a library
providing an API, I don't think there's a good reason to have micro
level releases. Probably it would be fine just to have integer releases,
but I think we may want to reserve a major version for when there are
larger breaks such as depending on new GNOME or Kolibri series.

This also fixes the project version declared to meson, which still had
the kolibri-installer-gnome version. When released, I expect to use
`vX.Y` tags.