ci: Run the CI flatpak build against the current branch
Rather than always against the latest `main`. This ensures that when
running CI on a PR, the code from the PR is what’s being built and
tested.

This uses the flatpak manifest support for pulling git repositories from
a local path, so it will always use the checkout containing the copy of
`org.endlessos.Key.Devel.json` which is being used.

Signed-off-by: Philip Withnall <philip@tecnocode.co.uk>