app: Allow XDG_CURRENT_DESKTOP with either capitalisation
Previously the desktop ID defined by Endless OS was `endless`. However,
we are now trying to standardise it, and it would be neater for it to be
`Endless`. Allow both cases of ID in the code, to allow for any
eventuality.

Strictly, this should be a case-sensitive comparison, but it seems
unlikely that someone else is going to claim `endless` in future.

See: https://gitlab.freedesktop.org/xdg/xdg-specs/-/merge_requests/73

Signed-off-by: Philip Withnall <philip@tecnocode.co.uk>

https://phabricator.endlessm.com/T34910