Build AppStream for metainfo validation
Since the metainfo validation was switched to using `appstreamcli` in
83878b3a493d93d29f661f0987fefac800e7adf0, it's not actually being run
during the flatpak build since `appstreamcli` isn't included in the
GNOME 44 runtime.

That will change with the GNOME 45 runtime, but in the meantime build
AppStream and it's dependencies during the flatpak build. Since these
are only test dependencies, all of their files can be cleaned up. Once
the flatpak is switched to the GNOME 45 runtime, this commit should be
reverted.