Default KOLIBRI_HOME to $XDG_HOME/kolibri
Upstream Kolibri defaults `KOLIBRI_HOME` to `~/.kolibri`, but in the
Flatpak context it's much better to use `$XDG_HOME/kolibri`
(`~/.local/share/kolibri` by default) so that the home directory doesn't
need to be exposed in the sandbox. Currently we're setting that via the
Flatpak finish args, but it's cleaner to handle that in the code as
suggested by a Flathub developer.