Initialize Kolibri debug from KOLIBRI_DEBUG environment variable
If the `KOLIBRI_DEBUG` environment variable is set to a truthy value,
pass `debug=True` to the Kolibri initialization routine. This enables
debug logging and the Django debug mode. This is roughly equivalent to
how `kolibri start` works since it treats `KOLIBRI_DEBUG` as equivalent
to the `--debug` CLI option.

Fixes: #123